+++ tests/test_aspectlib.py
@@ -1423,16 +1423,11 @@ def test_aspect_on_coroutine():
 def test_weave_module(strmod=None):
     calls = []
     from test_pkg1.test_pkg2 import test_mod
-    with aspectlib.weave(
-        strmod or test_mod,
-        lambda func: record(
-            callback=lambda inst, args, kwargs, name=func.__name__: calls.append((name, inst, args, kwargs))
-        )(func)
-    ):
+    with aspectlib.weave(strmod or test_mod, record(calls=calls, extended=True)):
         test_mod.target()
         obj = test_mod.Stuff()
         obj.meth()
-    assert calls == [('target', None, (), {}), ('meth', obj, (), {})]
+    assert calls == [(None, 'target', (), {}), (obj, 'meth', (), {})]
 
 def test_weave_module_as_str():
     test_weave_module("test_pkg1.test_pkg2.test_mod")